requirements.txt
flask>=2.3.0
pymysql>=1.1.0
werkzeug>=2.3.0
openai>=1.0.0
python-dotenv>=1.0.0
